+++ cmds/start-testnet-supernode.sh
@@ -1,4 +1,3 @@
 #!/bin/bash
 set -e
-symbol-bootstrap config -p testnet -a dual -t target/testnet-supernode -c test/supernode.yml -r --password 1111  $1 $2 $3
-symbol-bootstrap start --password 1111 --upgrade  -t target/testnet-supernode
+symbol-bootstrap start -p testnet -a dual -t target/testnet-supernode -c test/supernode.yml -r --password 1111  $1 $2 $3