+++ src/model/ConfigPreset.ts
@@ -248,6 +248,7 @@ export interface NodeConfigPreset {
     maxProofSize: number;
     maxTransactionsPerBlock: number;
     localNetworks: string;
+    rewardProgramAgentPort: number;
 }
 
 export interface NodePreset extends DockerServicePreset, Partial<NodeConfigPreset> {
@@ -274,6 +275,9 @@ export interface NodePreset extends DockerServicePreset, Partial<NodeConfigPrese
 
     vrfPrivateKey?: string;
     vrfPublicKey?: string;
+
+    agentPrivateKey?: string;
+    agentPublicKey?: string;
 
     //Broker specific
     brokerName?: string;
@@ -410,7 +414,7 @@ export interface CommonConfigPreset extends NodeConfigPreset, GatewayConfigPrese
     currencyMosaicId: string;
     harvestingMosaicId: string;
     baseNamespace: string;
-    rewardProgramControllerPublicKey?: string;
+    rewardProgramEnrollmentAddress?: string;
     networkType: NetworkType;
     //Nested Objects